Fix memory leak in dataset
The earlier version of the code opens all the HDF5 files at the
beginning and keeps them open throughout training. This makes the
dataset code easier to write, but allocates memory every time a
dataset is indexed into. This amounts to a memory leak that exhausts
memory in a matter of hours.

This commit instead opens an HDF5 file only when it is needed and
closes the file after getting the data. While this makes training
about 10% slower, this does stop the memory leak.

I also remove the preloading feature since it is not really used.

I also fix a bug where mean and variance are swapped when computing
the second moment for each dataset.